+++ gpt4all-backend/src/llmodel_shared.cpp
@@ -6,6 +6,7 @@
 #include <cstdint>
 #include <functional>
 #include <iostream>
+#include <iterator>
 #include <optional>
 #include <regex>
 #include <sstream>
@@ -66,18 +67,13 @@ void LLModel::prompt(const std::string &prompt,
         ss << "n_past=" << promptCtx.n_past << " is past end of context length=" << contextLength();
         throw std::out_of_range(ss.str());
     }
-    if (promptCtx.n_past > promptCtx.tokens.size()) {
+    if (promptCtx.n_past > inputLength()) {
         std::ostringstream ss;
-        ss << "n_past=" << promptCtx.n_past << " is past end of token cache length=" << promptCtx.tokens.size();
+        ss << "n_past=" << promptCtx.n_past << " is past end of token cache length=" << inputLength();
         throw std::out_of_range(ss.str());
     }
 
-    promptCtx.n_ctx = contextLength();
     promptCtx.n_batch = std::min(promptCtx.n_batch, LLMODEL_MAX_PROMPT_BATCH);
-
-    if (promptCtx.n_past < promptCtx.tokens.size())
-        promptCtx.tokens.resize(promptCtx.n_past);
-    m_tokenize_last_token = promptCtx.tokens.empty() ? -1 : promptCtx.tokens.back(); // not serialized
 
     // parse the prompt template
     std::vector<std::smatch> placeholders;
@@ -89,6 +85,8 @@ void LLModel::prompt(const std::string &prompt,
             return;
         }
     }
+
+    setTokenizeInputPosition(promptCtx.n_past);
 
     // tokenize the user prompt
     std::vector<Token> embd_inp;
@@ -118,7 +116,8 @@ void LLModel::prompt(const std::string &prompt,
     }
 
     // decode the user prompt
-    if (!decodePrompt(promptCallback, responseCallback, allowContextShift, promptCtx, embd_inp))
+    if (!decodePrompt(promptCallback, responseCallback, allowContextShift, promptCtx, embd_inp, /*isResponse*/ false,
+                      /*alwaysDecode*/ true))
         return; // error
 
     // decode the assistant's reply, either generated or spoofed
@@ -151,36 +150,67 @@ bool LLModel::decodePrompt(std::function<bool(int32_t)> promptCallback,
                            bool allowContextShift,
                            PromptContext &promptCtx,
                            std::vector<Token> embd_inp,
-                           bool isResponse) {
-    if ((int) embd_inp.size() > promptCtx.n_ctx - 4) {
+                           bool isResponse,
+                           bool alwaysDecode) {
+    if ((int) embd_inp.size() > contextLength() - 4) {
         // FIXME: (Adam) We should find a way to bubble these strings to the UI level to allow for
         // translation
         responseCallback(-1, "Your message was too long and could not be processed. Please try again with something shorter.");
         std::cerr << implementation().modelType() << " ERROR: The prompt is " << embd_inp.size() <<
-            " tokens and the context window is " << promptCtx.n_ctx << "!\n";
+            " tokens and the context window is " << contextLength() << "!\n";
         return false;
     }
 
     // FIXME(jared): There are mitigations for this situation, such as making room before
     // copying the prompt context, or restoring the KV cache when we restore the prompt
     // context.
-    if (!allowContextShift && promptCtx.n_past + embd_inp.size() > promptCtx.n_ctx) {
+    if (!allowContextShift && promptCtx.n_past + embd_inp.size() > contextLength()) {
         std::cerr << "LLModel Warning: Not enough space, n_past=" << promptCtx.n_past << ", n_eval=" << embd_inp.size()
-                  << ", n_ctx=" << promptCtx.n_ctx << "\n";
+                  << ", n_ctx=" << contextLength() << "\n";
         return false;
     }
 
-    // process the prompt in batches
+    // always decode something before generating, even if cached
+    if (alwaysDecode && embd_inp.empty()) {
+        auto cache = inputTokens();
+        if (!promptCtx.n_past)
+            throw std::runtime_error("zero token prompt is not supported");
+        assert(!cache.empty());
+        embd_inp.push_back(cache.back());
+        promptCtx.n_past--;
+    }
+
+    // Find the greatest n_past where the beginning of embd_inp matches the end of the token cache, starting at the
+    // requested n_past.
+    // This is used to skip unnecessary work when the prompt shares a common prefix with the previous result.
+    auto embd_inp_start = computeModelInputPosition(promptCtx, embd_inp);
+    size_t start_offset = embd_inp_start - embd_inp.begin();
+
+    // always decode up to a full batch before generating, even if cached
+    if (alwaysDecode)
+        start_offset -= std::min(promptCtx.n_batch, int32_t(start_offset));
+
+    setModelInputPosition(promptCtx, promptCtx.n_past + start_offset);
+
+    // execute the callback even for skipped tokens
     size_t i = 0;
+    for (; i < start_offset; i++) {
+        Token tok = embd_inp[i];
+        bool res = isResponse ? responseCallback(tok, tokenToString(tok)) : promptCallback(tok);
+        if (!res)
+            return false;
+    }
+
+    // process the prompt in batches
     while (i < embd_inp.size()) {
         size_t batch_end = std::min(i + promptCtx.n_batch, embd_inp.size());
-        std::vector<Token> batch(embd_inp.begin() + i, embd_inp.begin() + batch_end);
+        std::span<const Token> batch(embd_inp.begin() + i, embd_inp.begin() + batch_end);
 
         // Check if the context has run out...
-        if (promptCtx.n_past + int32_t(batch.size()) > promptCtx.n_ctx) {
+        if (promptCtx.n_past + int32_t(batch.size()) > contextLength()) {
             assert(allowContextShift);
             shiftContext(promptCtx);
-            assert(promptCtx.n_past + int32_t(batch.size()) <= promptCtx.n_ctx);
+            assert(promptCtx.n_past + int32_t(batch.size()) <= contextLength());
         }
 
         if (!evalTokens(promptCtx, batch)) {
@@ -190,9 +220,8 @@ bool LLModel::decodePrompt(std::function<bool(int32_t)> promptCallback,
 
         size_t tokens = batch_end - i;
         for (size_t t = 0; t < tokens; ++t) {
-            promptCtx.tokens.push_back(batch.at(t));
-            promptCtx.n_past += 1;
-            Token tok = batch.at(t);
+            Token tok = batch[t];
+            appendInputToken(promptCtx, tok);
             bool res = isResponse ? responseCallback(tok, tokenToString(tok)) : promptCallback(tok);
             if (!res)
                 return false;
@@ -232,8 +261,8 @@ void LLModel::generateResponse(std::function<bool(int32_t, const std::string&)> 
     // Don't even start if there is no room
     if (!promptCtx.n_predict)
         return;
-    if (!allowContextShift && promptCtx.n_past >= promptCtx.n_ctx) {
-        std::cerr << "LLModel Warning: Not enough space, n_past=" << promptCtx.n_past << ", n_ctx=" << promptCtx.n_ctx
+    if (!allowContextShift && promptCtx.n_past >= contextLength()) {
+        std::cerr << "LLModel Warning: Not enough space, n_past=" << promptCtx.n_past << ", n_ctx=" << contextLength()
                   << "\n";
         return;
     }
@@ -254,23 +283,22 @@ void LLModel::generateResponse(std::function<bool(int32_t, const std::string&)> 
 
         auto accept = [this, &promptCtx, &new_tok, allowContextShift]() -> bool {
             // Shift context if out of space
-            if (promptCtx.n_past >= promptCtx.n_ctx) {
+            if (promptCtx.n_past >= contextLength()) {
                 (void)allowContextShift;
                 assert(allowContextShift);
                 shiftContext(promptCtx);
-                assert(promptCtx.n_past < promptCtx.n_ctx);
+                assert(promptCtx.n_past < contextLength());
             }
 
             // Accept the token
             Token tok = std::exchange(new_tok, std::nullopt).value();
-            if (!evalTokens(promptCtx, { tok })) {
+            if (!evalTokens(promptCtx, { &tok, 1 })) {
                 // TODO(jared): raise an exception
                 std::cerr << implementation().modelType() << " ERROR: Failed to predict next token\n";
                 return false;
             }
 
-            promptCtx.tokens.push_back(tok);
-            promptCtx.n_past += 1;
+            appendInputToken(promptCtx, tok);
             return true;
         };
 
@@ -309,9 +337,9 @@ void LLModel::generateResponse(std::function<bool(int32_t, const std::string&)> 
         }
 
         // Optionally stop if the context will run out
-        if (!allowContextShift && promptCtx.n_past + cachedTokens.size() >= promptCtx.n_ctx) {
+        if (!allowContextShift && promptCtx.n_past + cachedTokens.size() >= contextLength()) {
             std::cerr << "LLModel Warning: Not enough space, n_past=" << promptCtx.n_past << ", n_ctx="
-                      << promptCtx.n_ctx << "\n";
+                      << contextLength() << "\n";
             stop = true;
         }
 
@@ -357,16 +385,17 @@ void LLModel::generateResponse(std::function<bool(int32_t, const std::string&)> 
         }
     }
 
-    auto &tokens = promptCtx.tokens;
-    if (tokens.size() < cachedTokens.size()) {
+    if (inputLength() < cachedTokens.size()) {
         /* This is theoretically possible if the longest stop sequence is greater than
          * n_ctx * contextErase tokens. */
         throw std::runtime_error("shifted too much context, can't go back");
     }
 
-    auto discard_start = tokens.end() - cachedTokens.size();
-    assert(std::equal(discard_start, tokens.end(), cachedTokens.begin()));
-    tokens.erase(discard_start, tokens.end());
+#ifndef NDEBUG
+    auto inp = inputTokens();
+    auto discard_start = inp.end() - cachedTokens.size();
+    assert(std::equal(discard_start, inp.end(), cachedTokens.begin()));
+#endif
 
     promptCtx.n_past -= cachedTokens.size();
 }