+++ gpt4all-chat/src/chatapi.cpp
@@ -51,7 +51,6 @@ bool ChatAPI::loadModel(const std::string &modelPath, int n_ctx, int ngl)
 void ChatAPI::setThreadCount(int32_t n_threads)
 {
     Q_UNUSED(n_threads);
-    qt_noop();
 }
 
 int32_t ChatAPI::threadCount() const
@@ -66,24 +65,6 @@ ChatAPI::~ChatAPI()
 bool ChatAPI::isModelLoaded() const
 {
     return true;
-}
-
-// All three of the state virtual functions are handled custom inside of chatllm save/restore
-size_t ChatAPI::stateSize() const
-{
-    throw std::logic_error("not implemented");
-}
-
-size_t ChatAPI::saveState(std::span<uint8_t> dest) const
-{
-    Q_UNUSED(dest);
-    throw std::logic_error("not implemented");
-}
-
-size_t ChatAPI::restoreState(std::span<const uint8_t> src)
-{
-    Q_UNUSED(src);
-    throw std::logic_error("not implemented");
 }
 
 void ChatAPI::prompt(const std::string &prompt,